Fixed compilation on MSVC 2010, added MSVC 2010 project

+++ src/gusanos/gui/detail/gss.cpp
@@ -25,7 +25,7 @@ struct GSSImpl : public TGrammar<GSSImpl>
 	size_t read(char* p, size_t s)
 	{
 		str.read(p, s);
-		return str.gcount();
+		return (size_t)str.gcount();
 	}
 	
 	void reportError(std::string const& error, Location loc)
