+++ src/gusanos/omfgscript/omfg_script.cpp
@@ -582,7 +582,7 @@ struct ParserImpl : public TGrammar<ParserImpl>
 	size_t read(char* p, size_t s)
 	{
 		str.read(p, s);
-		return str.gcount();
+		return (size_t)str.gcount();
 	}
 	
 	void reportError(std::string const& error, Location loc)
